treadmill/requirements.txt
flask
flask-cors
RPi.GPIO
picamera
pyserial